requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
comtypes==1.1.14
distlib==0.3.6
filelock==3.8.0
idna==3.3
joblib==1.1.0
numpy==1.23.2
platformdirs==2.5.2
PyAudio==0.2.12
pypiwin32==223
python-dotenv==0.21.0
pyttsx3==2.90
pywin32==304
requests==2.28.1
scipy==1.9.1
SpeechRecognition==3.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
urllib3==1.26.12
virtualenv==20.16.4
